fix(utils): fix error when the text "@@" appears twice within a single line of code
precise-commits could mistake part of the source code for a
unified diff hunk marker if the text "@@" appears twice in
one line.

In unified diff format, markers always appear in the first
column and the text of the diff always appears from the
second column onwards, so ignore anything that looks like
a hunk marker if it does not start in the first column.